+++ .check.exs
@@ -1,15 +1,10 @@
+# SPDX-FileCopyrightText: 2024 splode contributors <https://github.com/ash-project/splode/graphs/contributors>
+#
+# SPDX-License-Identifier: MIT
+
 [
-  ## all available options with default values (see `mix check` docs for description)
-  # parallel: true,
-  # skipped: true,
-
-  ## list of tools (see `mix check` docs for defaults)
   tools: [
-    ## curated tools may be disabled (e.g. the check for compilation warnings)
-    # {:compiler, false},
-
-    ## ...or adjusted (e.g. use one-line formatter for more compact credo output)
-    # {:credo, "mix credo --format oneline"},
-    {:doctor, false}
+    {:doctor, false},
+    {:reuse, command: ["pipx", "run", "reuse", "lint", "-q"]}
   ]
 ]