+++ .doctor.exs
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2024 splode contributors <https://github.com/ash-project/splode/graphs/contributors>
+#
+# SPDX-License-Identifier: MIT
+
 %Doctor.Config{
   ignore_modules: [ ],
   ignore_paths: [],