+++ lib/splode/stacktrace.ex
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2024 splode contributors <https://github.com/ash-project/splode/graphs/contributors>
+#
+# SPDX-License-Identifier: MIT
+
 defmodule Splode.Stacktrace do
   @moduledoc "A placeholder for a stacktrace so that we can avoid printing it everywhere"
   defstruct [:stacktrace]