+++ test/test_helper.exs
@@ -1,1 +1,5 @@
+# SPDX-FileCopyrightText: 2024 splode contributors <https://github.com/ash-project/splode/graphs/contributors>
+#
+# SPDX-License-Identifier: MIT
+
 ExUnit.start()